backend/requirements.txt
# Core FastAPI dependencies (using exact versions from original)
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Flask (for gradual migration from original)
flask==3.0.0
flask-cors==4.0.0

# AI/ML dependencies (exact versions from original SkillsProfGen)
openai==1.51.0
anthropic==0.36.0
google-generativeai==0.8.0
langchain==0.3.0
langchain-community==0.3.0
langchain-anthropic==0.2.0
langchain-openai==0.2.0

# Vector store and embeddings (exact versions)
chromadb==0.5.3
torch>=2.6.0
sentence-transformers==3.0.1
numpy==1.26.4

# API and data handling (exact versions)
requests==2.32.3
pandas==2.2.2
sqlalchemy==2.0.35

# Additional utilities
aiofiles==24.1.0
httpx==0.27.0

# Testing dependencies
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
coverage[toml]==7.6.1

# Development tools
black==24.8.0
isort==5.13.2
mypy==1.11.2
pre-commit==3.8.0
watchdog==5.0.3

# Additional dependencies for SkillsProfGen functionality
transformers>=4.30.0
faiss-cpu>=1.7.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
loguru>=0.7.0
python-dateutil>=2.8.2